requirements.txt
numpy
pandas
matplotlib
statsmodels
tiingo
scikit-learn
tensorflow
dotenv
pyarrow
duckdb
alpaca-trade-api
alpaca-py>=0.10.0

